finished implementation of extractor program
