+++ extractor/src/main/java/co/adeshina/c19terma/extractor/exception/ConfigPropertiesInitException.java
@@ -1,0 +1,8 @@
+package co.adeshina.c19terma.extractor.exception;
+
+public class ConfigPropertiesInitException extends Exception {
+
+    public ConfigPropertiesInitException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
