+++ extractor/src/main/java/co/adeshina/c19terma/extractor/http/dto/UserDto.java
@@ -1,0 +1,26 @@
+package co.adeshina.c19terma.extractor.http.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class UserDto {
+
+    private String id;
+    private boolean verified;
+
+    public boolean isVerified() {
+        return verified;
+    }
+
+    public void setVerified(boolean verified) {
+        this.verified = verified;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+}
