+++ extractor/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+# Twitter API Credentials --- TODO create profile for project before testing
+
+twitter.consumer.secret=${consumer.secret}
+twitter.consumer.key=${consumer.key}
+
+# Kafka Config
+
+kafka.acks=${}
+kafka.bootstrap.servers=${}
+kafka.client.id=${}
+kafka.topic=${}
