+++ extractor/src/test/resources/test.properties
@@ -1,0 +1,12 @@
+
+# Twitter Config
+
+twitter.secret=secret1234
+twitter.key=key1234
+
+# Kafka Config
+
+kafka.acks=acks1234
+kafka.server=server1234
+kafka.client.id=id1234
+kafka.topic=topic1234
