--- stream_rules.json
@@ -1,17 +1,0 @@
-{
-  "add": [
-    {"value":  ""},
-    {"value":  ""},
-    {"value":  ""},
-    {"value":  ""},
-    {"value":  ""}
-  ]
-}
-
-
-
-{
-  "delete": {
-    "ids": ids.toString();
-  }
-}
