--- tweet-extractor/pom.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>co.adeshina.c19terma</groupId>
-    <artifactId>c19terma-parent</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>tweet-extractor</artifactId>
-  <name>Tweet Extractor</name>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>co.adeshina.c19terma</groupId>
-      <artifactId>common</artifactId>
-      <version>${parent.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>com.squareup.okhttp3</groupId>
-      <artifactId>okhttp</artifactId>
-    </dependency>
-  </dependencies>
-
- <build>
-   <plugins>
-     <plugin>
-       <groupId>org.apache.maven.plugins</groupId>
-       <artifactId>maven-compiler-plugin</artifactId>
-     </plugin>
-
-     <plugin>
-       <groupId>org.apache.maven.plugins</groupId>
-       <artifactId>maven-surefire-plugin</artifactId>
-     </plugin>
-
-     <plugin>
-       <groupId>org.apache.maven.plugins</groupId>
-       <artifactId>maven-shade-plugin</artifactId>
-     </plugin>
-   </plugins>
-
-   <resources>
-     <resource>
-       <directory>src/main/resources</directory>
-       <filtering>true</filtering>
-     </resource>
-   </resources>
- </build>
-
-</project>
