--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/exception/ApiClientException.java
@@ -1,14 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.exception;
-
-public class ApiClientException extends Exception {
-
-    public ApiClientException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ApiClientException(String message) {
-        super(message);
-    }
-
-    // todo: extend exception or Runtime Exception??
-}
