--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/exception/TweetExtractorException.java
@@ -1,9 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.exception;
-
-// todo: is this in use??
-public class TweetExtractorException extends Exception {
-
-    public TweetExtractorException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
