--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/http/TwitterUserApiClient.java
@@ -1,17 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.http;
-
-import co.adeshina.c19terma.tweetextractor.exception.ApiClientException;
-import co.adeshina.c19terma.tweetextractor.http.dto.UserDto;
-
-/**
- *
- */
-public interface TwitterUserApiClient {
-
-    /**
-     *
-     * @param id
-     * @return
-     */
-    UserDto findUser(String id) throws ApiClientException;
-}
