--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/http/dto/BearerTokenDto.java
@@ -1,30 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.http.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class BearerTokenDto {
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    public String getTokenType() {
-        return tokenType;
-    }
-
-    public void setTokenType(String tokenType) {
-        this.tokenType = tokenType;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-}
