--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/http/dto/RulesDto.java
@@ -1,42 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.http.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.util.List;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class RulesDto {
-
-    private List<Rule> data;
-
-    public List<Rule> getData() {
-        return data;
-    }
-
-    public void setData(List<Rule> data) {
-        this.data = data;
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static class Rule {
-
-        private String value;
-        private String id;
-
-        public String getValue() {
-            return value;
-        }
-
-        public void setValue(String value) {
-            this.value = value;
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public void setId(String id) {
-            this.id = id;
-        }
-    }
-
-}
