--- tweet-extractor/src/main/java/co/adeshina/c19terma/tweetextractor/http/dto/TweetDto.java
@@ -1,29 +1,0 @@
-package co.adeshina.c19terma.tweetextractor.http.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class TweetDto {
-
-    private String text;
-
-    @JsonProperty("author_id")
-    private String authorId;
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public String getAuthorId() {
-        return authorId;
-    }
-
-    public void setAuthorId(String authorId) {
-        this.authorId = authorId;
-    }
-}
