--- tweet-extractor/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-# Twitter API Credentials --- TODO create profile for project before testing
-
-consumer.secret=${consumer.secret}
-consumer.key=${consumer.key}
-
-# Kafka Config
